# openapi.json
{"openapi":"3.1.0","info":{"title":"IBEX35 RAG API","description":"Production-ready Retrieval-Augmented Generation system for IBEX35 financial reports. Powered by LlamaIndex + Ollama + ChromaDB.","version":"0.1.0"},"paths":{"/metrics":{"get":{"summary":"Metrics","description":"Endpoint that serves Prometheus metrics.","operationId":"metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health Check","description":"Full health check: vector store, Redis, Ollama.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/health/live":{"get":{"tags":["health"],"summary":"Liveness","description":"Kubernetes liveness probe — just confirms the process is alive.","operationId":"liveness_health_live_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Liveness Health Live Get"}}}}}}},"/health/ready":{"get":{"tags":["health"],"summary":"Readiness","description":"Kubernetes readiness probe — confirms the app can serve traffic.","operationId":"readiness_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Readiness Health Ready Get"}}}}}}},"/api/v1/query":{"post":{"tags":["query"],"summary":"Query Rag","description":"Query the RAG system with a financial question.\n\n- Checks Redis cache first (SHA-256 keyed by question + filter).\n- Falls back to LlamaIndex retrieval + Ollama generation.\n- Rate limited per IP (sliding window via Redis).","operationId":"query_rag_api_v1_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/agent":{"post":{"tags":["query"],"summary":"Query Agent","description":"Query the ReAct financial agent.\n\nCombines RAG knowledge base with real-time market tools (yfinance).\nBest for complex, multi-step questions that require both fundamentals and market data.","operationId":"query_agent_api_v1_agent_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/ingest":{"post":{"tags":["ingestion"],"summary":"Trigger Ingestion","description":"Trigger PDF ingestion in the background.\n\nLoads all PDFs from the configured directory, chunks them, computes embeddings\nwith Ollama, and stores vectors in ChromaDB. Tracks the run in MLflow.","operationId":"trigger_ingestion_api_v1_ingest_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestionRequest"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/IngestionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AgentResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"latency_seconds":{"type":"number","title":"Latency Seconds"},"steps_taken":{"type":"integer","title":"Steps Taken"},"tools_used":{"items":{"type":"string"},"type":"array","title":"Tools Used"},"query":{"type":"string","title":"Query"}},"type":"object","required":["answer","latency_seconds","steps_taken","tools_used","query"],"title":"AgentResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"version":{"type":"string","title":"Version"},"vector_store":{"type":"string","title":"Vector Store"},"redis":{"type":"string","title":"Redis"},"ollama":{"type":"string","title":"Ollama"},"collection_count":{"type":"integer","title":"Collection Count"}},"type":"object","required":["status","version","vector_store","redis","ollama","collection_count"],"title":"HealthResponse"},"IngestionRequest":{"properties":{"pdf_dir":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Dir","description":"Override default PDF directory path"}},"type":"object","title":"IngestionRequest"},"IngestionResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"total_documents":{"type":"integer","title":"Total Documents"},"total_nodes":{"type":"integer","title":"Total Nodes"},"total_companies":{"type":"integer","title":"Total Companies"},"duration_seconds":{"type":"number","title":"Duration Seconds"},"companies":{"items":{"type":"string"},"type":"array","title":"Companies"},"errors":{"items":{"type":"string"},"type":"array","title":"Errors"}},"type":"object","required":["success","total_documents","total_nodes","total_companies","duration_seconds","companies"],"title":"IngestionResponse"},"QueryRequest":{"properties":{"question":{"type":"string","maxLength":1000,"minLength":5,"title":"Question","description":"Financial question"},"company_filter":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Filter","description":"Filter results to a specific IBEX35 company (e.g. 'IBERDROLA')"},"use_agent":{"type":"boolean","title":"Use Agent","description":"Use the ReAct agent (combines RAG + real-time data). Slower but more capable.","default":false}},"type":"object","required":["question"],"title":"QueryRequest"},"QueryResponse":{"properties":{"answer":{"type":"string","title":"Answer"},"sources":{"items":{"$ref":"#/components/schemas/SourceDoc"},"type":"array","title":"Sources"},"latency_seconds":{"type":"number","title":"Latency Seconds"},"from_cache":{"type":"boolean","title":"From Cache","default":false},"query":{"type":"string","title":"Query"}},"type":"object","required":["answer","latency_seconds","query"],"title":"QueryResponse"},"SourceDoc":{"properties":{"company":{"type":"string","title":"Company"},"ticker":{"type":"string","title":"Ticker"},"score":{"type":"number","title":"Score"}},"type":"object","required":["company","ticker","score"],"title":"SourceDoc"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}